requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
da-vinci==0.5.0
diff-match-patch==20241021
Django==5.1.7
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-import-export==4.3.7
django-js-asset==3.1.2
django-rosetta==0.10.2
django-simple-history==3.8.0
django-thumbnails==0.8.0
idna==3.10
pillow==11.1.0
polib==1.2.0
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.3
tablib==3.8.0
urllib3==2.3.0
django-modeltranslation==0.19.13